start moving BOM methods to BOM class
