+++ dependencies/src/main/java/com/google/cloud/tools/opensource/dependencies/Bom.java
@@ -19,9 +19,15 @@ package com.google.cloud.tools.opensource.dependencies;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableSet;
 import org.eclipse.aether.artifact.Artifact;
+import org.eclipse.aether.artifact.ArtifactProperties;
 
 public final class Bom {
+  
+  private static final ImmutableSet<String> BOM_SKIP_ARTIFACT_IDS =
+      ImmutableSet.of("google-cloud-logging-logback", "google-cloud-contrib");
+
   
   private final ImmutableList<Artifact> artifacts;
   private final String coordinates;
@@ -51,5 +57,25 @@ public final class Bom {
    */
   public String getCoordinates() {
     return coordinates;
+  }
+
+  /** Returns true if the {@code artifact} in BOM should be skipped for checks. */
+  public static boolean shouldSkipBomMember(Artifact artifact) {
+    if ("testlib".equals(artifact.getClassifier())) {
+      // we don't report on test libraries
+      return true;
+    }
+  
+    String type = artifact.getProperty(ArtifactProperties.TYPE, "jar");
+    if ("test-jar".equals(type)) {
+      return true;
+    }
+  
+    // TODO remove this hack once we get these out of google-cloud-java's BOM
+    if (BOM_SKIP_ARTIFACT_IDS.contains(artifact.getArtifactId())) {
+      return true;
+    }
+  
+    return false;
   }
 }
