+++ enforcer-rules/src/main/java/com/google/cloud/tools/dependencies/enforcer/LinkageCheckerRule.java
@@ -29,6 +29,8 @@ import com.google.cloud.tools.opensource.classpath.ClassReferenceGraph;
 import com.google.cloud.tools.opensource.classpath.LinkageChecker;
 import com.google.cloud.tools.opensource.classpath.SymbolProblem;
 import com.google.cloud.tools.opensource.dependencies.ArtifactProblem;
+import com.google.cloud.tools.opensource.dependencies.Bom;
+import com.google.cloud.tools.opensource.dependencies.Bom.shouldSkipBomMember;
 import com.google.cloud.tools.opensource.dependencies.DependencyGraph;
 import com.google.cloud.tools.opensource.dependencies.DependencyGraphBuilder;
 import com.google.cloud.tools.opensource.dependencies.DependencyPath;
@@ -366,7 +368,7 @@ public class LinkageCheckerRule extends AbstractNonCacheableEnforcerRule {
         bomProject.getDependencyManagement().getDependencies().stream()
             .map(dependency -> RepositoryUtils.toDependency(dependency, artifactTypeRegistry))
             .map(Dependency::getArtifact)
-            .filter(artifact -> !shouldSkipBomMember(artifact))
+            .filter(artifact -> !Bom.shouldSkipBomMember(artifact))
             .collect(toImmutableList());
 
     ClassPathResult result = classPathBuilder.resolve(artifacts);
